requirements.txt
streamlit
openai
chromadb
duckdb
requests
beautifulsoup4
pytest
python-dotenv
tqdm
urllib3
python-dotenv
